Refactored runner to live in cucumber package
