+++ src/main/g8/run_qa.sh
@@ -3,4 +3,4 @@ ENV="qa"
 BROWSER="chrome"
 DRIVER_PATH=/usr/local/bin/chromedriver
 
-sbt -Dlogback.configurationFile=logback.xml -Dbrowser=\$BROWSER -Denvironment=\$ENV -Dwebdriver.chrome.driver=\${DRIVER_PATH} 'test-only uk.gov.hmrc.test.ui.runner.RunnerQA'
+sbt -Dlogback.configurationFile=logback.xml -Dbrowser=\$BROWSER -Denvironment=\$ENV -Dwebdriver.chrome.driver=\${DRIVER_PATH} 'test-only uk.gov.hmrc.test.ui.cucumber.runner.RunnerQA'
