+++ src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/driver/ProxySupport.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.hmrc.test.ui.runner.driver
+package uk.gov.hmrc.test.ui.driver
 
 import java.net.InetSocketAddress
