+++ src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/driver/browsers/FirefoxBrowser.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.hmrc.test.ui.runner.driver.browsers
+package uk.gov.hmrc.test.ui.driver.browsers
 
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.firefox.{FirefoxDriver, FirefoxOptions, FirefoxProfile}
