+++ src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/pages/BasePage.scala
@@ -1,20 +1,4 @@
 /*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.gov.hmrc.test.ui.pages
-/*
  * Copyright 2018 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -29,6 +13,8 @@ package uk.gov.hmrc.test.ui.pages
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package uk.gov.hmrc.test.ui.pages
+
 import java.util.concurrent.TimeUnit
 
 import cucumber.api.DataTable
@@ -36,7 +22,7 @@ import org.junit.Assert
 import org.openqa.selenium.support.ui.{ExpectedConditions, FluentWait, Select}
 import org.openqa.selenium.{By, NoAlertPresentException, WebDriver}
 import org.scalatest.Matchers
-import uk.gov.hmrc.test.ui.runner.driver.Driver
+import uk.gov.hmrc.test.ui.driver.Driver
 import uk.gov.hmrc.test.ui.conf.TestConfiguration.getProperty
 
 object BasePage extends BasePage {
